xDai default values + add EthMinGasPriceDefault option (more xDai compat) (#4421)
* xDai defaults; add minimum gas price
