+++ core/services/gasupdater/gas_updater.go
@@ -64,8 +64,10 @@ type (
 		GasUpdaterTransactionPercentile() uint16
 		GasUpdaterBatchSize() uint32
 		EthMaxGasPriceWei() *big.Int
+		EthMinGasPriceWei() *big.Int
 		EthFinalityDepth() uint
 		SetEthGasPriceDefault(value *big.Int) error
+		ChainID() *big.Int
 	}
 
 	gasUpdater struct {
@@ -192,7 +194,9 @@ func (gu *gasUpdater) Recalculate(head models.Head) {
 		logger.Warnw("GasUpdater: cannot calculate percentile gas price", "err", err)
 		return
 	}
-	gasPriceGwei := fmt.Sprintf("%.2f", float64(percentileGasPrice)/1000000000)
+	float := new(big.Float).SetInt(percentileGasPrice)
+	gwei, _ := big.NewFloat(0).Quo(float, big.NewFloat(1000000000)).Float64()
+	gasPriceGwei := fmt.Sprintf("%.2f", gwei)
 
 	var numsInHistory []int64
 	for _, b := range gu.rollingBlockHistory {
@@ -209,7 +213,7 @@ func (gu *gasUpdater) Recalculate(head models.Head) {
 		gu.logger.Errorw("GasUpdater: error setting gas price", "err", err)
 		return
 	}
-	promGasUpdaterSetGasPrice.WithLabelValues(fmt.Sprintf("%v%%", percentile)).Set(float64(percentileGasPrice))
+	promGasUpdaterSetGasPrice.WithLabelValues(fmt.Sprintf("%v%%", percentile)).Set(float64(percentileGasPrice.Int64()))
 }
 
 func (gu *gasUpdater) FetchBlocks(ctx context.Context, head models.Head) error {
@@ -329,39 +333,51 @@ func (gu *gasUpdater) batchFetch(ctx context.Context, reqs []rpc.BatchElem) erro
 	return nil
 }
 
-func (gu *gasUpdater) percentileGasPrice(percentile int) (int64, error) {
-	gasPrices := make([]int64, 0)
+func (gu *gasUpdater) percentileGasPrice(percentile int) (*big.Int, error) {
+	minGasPriceWei := gu.config.EthMinGasPriceWei()
+	chainID := gu.config.ChainID()
+	gasPrices := make([]*big.Int, 0)
 	for _, block := range gu.rollingBlockHistory {
 		for _, tx := range block.Transactions {
-			// GasLimit 0 is impossible on Ethereum official, but IS possible
-			// on forks/clones such as RSK. We should ignore these transactions
-			// if they come up since they are not normal.
-			if tx.GasLimit > 0 {
-				gasPrices = append(gasPrices, tx.GasPrice.Int64())
+			if isUsableTx(tx, minGasPriceWei, chainID) {
+				gasPrices = append(gasPrices, tx.GasPrice)
 			}
 		}
 	}
 	if len(gasPrices) == 0 {
-		return 0, errors.New("no suitable transactions")
+		return big.NewInt(0), errors.New("no suitable transactions")
 	}
-	sort.Slice(gasPrices, func(i, j int) bool { return gasPrices[i] < gasPrices[j] })
+	sort.Slice(gasPrices, func(i, j int) bool { return gasPrices[i].Cmp(gasPrices[j]) < 0 })
 	idx := ((len(gasPrices) - 1) * percentile) / 100
 	for i := 0; i <= 100; i += 5 {
 		jdx := ((len(gasPrices) - 1) * i) / 100
-		promGasUpdaterAllPercentiles.WithLabelValues(fmt.Sprintf("%v%%", i)).Set(float64(gasPrices[jdx]))
+		promGasUpdaterAllPercentiles.WithLabelValues(fmt.Sprintf("%v%%", i)).Set(float64(gasPrices[jdx].Int64()))
 	}
 	return gasPrices[idx], nil
 }
 
-func (gu *gasUpdater) setPercentileGasPrice(gasPrice int64) error {
-	bigGasPrice := big.NewInt(gasPrice)
-	if bigGasPrice.Cmp(gu.config.EthMaxGasPriceWei()) > 0 {
-		gu.logger.Warnw(fmt.Sprintf("Calculated gas price of %s Wei exceeds ETH_MAX_GAS_PRICE_WEI=%[2]s, setting gas price to the maximum allowed value of %[2]s Wei instead", bigGasPrice.String(), gu.config.EthMaxGasPriceWei().String()), "gasPriceWei", bigGasPrice, "maxGasPriceWei", gu.config.EthMaxGasPriceWei())
+func (gu *gasUpdater) setPercentileGasPrice(gasPrice *big.Int) error {
+	if gasPrice.Cmp(gu.config.EthMaxGasPriceWei()) > 0 {
+		gu.logger.Warnw(fmt.Sprintf("Calculated gas price of %s Wei exceeds ETH_MAX_GAS_PRICE_WEI=%[2]s, setting gas price to the maximum allowed value of %[2]s Wei instead", gasPrice.String(), gu.config.EthMaxGasPriceWei().String()), "gasPriceWei", gasPrice, "maxGasPriceWei", gu.config.EthMaxGasPriceWei())
 		return gu.config.SetEthGasPriceDefault(gu.config.EthMaxGasPriceWei())
 	}
-	return gu.config.SetEthGasPriceDefault(bigGasPrice)
+	if gasPrice.Cmp(gu.config.EthMinGasPriceWei()) < 0 {
+		gu.logger.Warnw(fmt.Sprintf("Calculated gas price of %s Wei falls below ETH_MIN_GAS_PRICE_WEI=%[2]s, setting gas price to the minimum allowed value of %[2]s Wei instead", gasPrice.String(), gu.config.EthMaxGasPriceWei().String()), "gasPriceWei", gasPrice, "maxGasPriceWei", gu.config.EthMaxGasPriceWei())
+		return gu.config.SetEthGasPriceDefault(gu.config.EthMinGasPriceWei())
+	}
+	return gu.config.SetEthGasPriceDefault(gasPrice)
 }
 
 func (gu *gasUpdater) RollingBlockHistory() []Block {
 	return gu.rollingBlockHistory
+}
+
+func isUsableTx(tx Transaction, minGasPriceWei, chainID *big.Int) bool {
+	// GasLimit 0 is impossible on Ethereum official, but IS possible
+	// on forks/clones such as RSK. We should ignore these transactions
+	// if they come up on any chain since they are not normal.
+	if tx.GasLimit == 0 {
+		return false
+	}
+	return chainSpecificIsUsableTx(tx, minGasPriceWei, chainID)
 }
