+++ core/services/gasupdater/gas_updater_test.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"math"
 	"math/big"
 	"reflect"
 	"testing"
@@ -31,12 +32,15 @@ func TestGasUpdater_Start(t *testing.T) {
 	var historySize uint16 = 2
 	var ethFinalityDepth uint = 42
 	var percentile uint16 = 35
+	minGasPrice := big.NewInt(1)
 
 	config.On("GasUpdaterBatchSize").Return(batchSize)
 	config.On("GasUpdaterBlockDelay").Return(blockDelay)
 	config.On("GasUpdaterBlockHistorySize").Return(historySize)
 	config.On("EthFinalityDepth").Return(ethFinalityDepth)
 	config.On("GasUpdaterTransactionPercentile").Return(percentile)
+	config.On("EthMinGasPriceWei").Return(minGasPrice)
+	config.On("ChainID").Return(big.NewInt(0))
 
 	t.Run("loads initial state", func(t *testing.T) {
 		ethClient := new(mocks.Client)
@@ -116,6 +120,7 @@ func TestGasUpdater_FetchBlocks(t *testing.T) {
 
 		var blockDelay uint16 = 3
 		var historySize uint16 = 0
+		config.On("ChainID").Return(big.NewInt(0))
 		config.On("GasUpdaterBlockDelay").Return(blockDelay)
 		config.On("GasUpdaterBlockHistorySize").Return(historySize)
 
@@ -276,7 +281,9 @@ func TestGasUpdater_FetchBlocksAndRecalculate(t *testing.T) {
 	config.On("GasUpdaterTransactionPercentile").Return(uint16(35))
 	config.On("GasUpdaterBlockHistorySize").Return(uint16(3))
 	config.On("EthMaxGasPriceWei").Return(big.NewInt(1000))
+	config.On("EthMinGasPriceWei").Return(big.NewInt(0))
 	config.On("GasUpdaterBatchSize").Return(uint32(0))
+	config.On("ChainID").Return(big.NewInt(0))
 
 	guIface := gasupdater.NewGasUpdater(ethClient, config)
 	gu := gasupdater.GasUpdaterToStruct(guIface)
@@ -323,12 +330,15 @@ func TestGasUpdater_Recalculate(t *testing.T) {
 	t.Parallel()
 
 	maxGasPrice := big.NewInt(100)
+	minGasPrice := big.NewInt(10)
 
 	t.Run("does not crash or set gas price to zero if there are no transactions", func(t *testing.T) {
 		ethClient := new(mocks.Client)
 		config := new(gumocks.Config)
 
 		config.On("GasUpdaterTransactionPercentile").Return(uint16(35))
+		config.On("EthMinGasPriceWei").Return(big.NewInt(1))
+		config.On("ChainID").Return(big.NewInt(0))
 
 		guIface := gasupdater.NewGasUpdater(ethClient, config)
 		gu := gasupdater.GasUpdaterToStruct(guIface)
@@ -354,7 +364,9 @@ func TestGasUpdater_Recalculate(t *testing.T) {
 		config := new(gumocks.Config)
 
 		config.On("EthMaxGasPriceWei").Return(maxGasPrice)
+		config.On("EthMinGasPriceWei").Return(minGasPrice)
 		config.On("GasUpdaterTransactionPercentile").Return(uint16(35))
+		config.On("ChainID").Return(big.NewInt(0))
 
 		guIface := gasupdater.NewGasUpdater(ethClient, config)
 		gu := gasupdater.GasUpdaterToStruct(guIface)
@@ -381,12 +393,48 @@ func TestGasUpdater_Recalculate(t *testing.T) {
 		config.AssertExpectations(t)
 	})
 
+	t.Run("sets gas price to ETH_MIN_GAS_PRICE_WEI if the calculation would otherwise fall below it", func(t *testing.T) {
+		ethClient := new(mocks.Client)
+		config := new(gumocks.Config)
+
+		config.On("EthMaxGasPriceWei").Return(maxGasPrice)
+		config.On("EthMinGasPriceWei").Return(minGasPrice)
+		config.On("GasUpdaterTransactionPercentile").Return(uint16(35))
+		config.On("ChainID").Return(big.NewInt(0))
+
+		guIface := gasupdater.NewGasUpdater(ethClient, config)
+		gu := gasupdater.GasUpdaterToStruct(guIface)
+
+		blocks := []gasupdater.Block{
+			gasupdater.Block{
+				Number:       0,
+				Hash:         cltest.NewHash(),
+				Transactions: cltest.TransactionsFromGasPrices(5),
+			},
+			gasupdater.Block{
+				Number:       1,
+				Hash:         cltest.NewHash(),
+				Transactions: cltest.TransactionsFromGasPrices(7),
+			},
+		}
+
+		config.On("SetEthGasPriceDefault", minGasPrice).Return(nil)
+		gasupdater.SetRollingBlockHistory(gu, blocks)
+
+		gu.Recalculate(*cltest.Head(1))
+
+		ethClient.AssertExpectations(t)
+		config.AssertExpectations(t)
+	})
+
 	t.Run("ignores any transaction with a zero gas limit", func(t *testing.T) {
 		ethClient := new(mocks.Client)
 		config := new(gumocks.Config)
 
 		config.On("EthMaxGasPriceWei").Return(maxGasPrice)
+		config.On("EthMinGasPriceWei").Return(minGasPrice)
 		config.On("GasUpdaterTransactionPercentile").Return(uint16(100))
+		config.On("ChainID").Return(big.NewInt(0))
 
 		guIface := gasupdater.NewGasUpdater(ethClient, config)
 		gu := gasupdater.GasUpdaterToStruct(guIface)
@@ -424,6 +472,118 @@ func TestGasUpdater_Recalculate(t *testing.T) {
 		config.AssertExpectations(t)
 	})
 
+	t.Run("takes into account zero priced transctions if chain is not xDai", func(t *testing.T) {
+		// Because everyone loves free gas!
+		ethClient := new(mocks.Client)
+		config := new(gumocks.Config)
+
+		config.On("EthMaxGasPriceWei").Return(maxGasPrice)
+		config.On("EthMinGasPriceWei").Return(big.NewInt(0))
+		config.On("GasUpdaterTransactionPercentile").Return(uint16(50))
+		config.On("ChainID").Return(big.NewInt(0))
+
+		guIface := gasupdater.NewGasUpdater(ethClient, config)
+		gu := gasupdater.GasUpdaterToStruct(guIface)
+
+		b1Hash := cltest.NewHash()
+
+		blocks := []gasupdater.Block{
+			gasupdater.Block{
+				Number:       0,
+				Hash:         b1Hash,
+				ParentHash:   common.Hash{},
+				Transactions: cltest.TransactionsFromGasPrices(0, 0, 0, 0, 100),
+			},
+		}
+
+		config.On("SetEthGasPriceDefault", big.NewInt(0)).Return(nil)
+		gasupdater.SetRollingBlockHistory(gu, blocks)
+
+		gu.Recalculate(*cltest.Head(0))
+
+		ethClient.AssertExpectations(t)
+		config.AssertExpectations(t)
+	})
+
+	t.Run("ignores zero priced transactions on xDai", func(t *testing.T) {
+		ethClient := new(mocks.Client)
+		config := new(gumocks.Config)
+
+		config.On("EthMaxGasPriceWei").Return(maxGasPrice)
+		config.On("EthMinGasPriceWei").Return(big.NewInt(100))
+		config.On("GasUpdaterTransactionPercentile").Return(uint16(50))
+		config.On("ChainID").Return(big.NewInt(100))
+
+		guIface := gasupdater.NewGasUpdater(ethClient, config)
+		gu := gasupdater.GasUpdaterToStruct(guIface)
+
+		b1Hash := cltest.NewHash()
+
+		blocks := []gasupdater.Block{
+			gasupdater.Block{
+				Number:       0,
+				Hash:         b1Hash,
+				ParentHash:   common.Hash{},
+				Transactions: cltest.TransactionsFromGasPrices(0, 0, 0, 0, 100),
+			},
+		}
+
+		config.On("SetEthGasPriceDefault", big.NewInt(100)).Return(nil)
+		gasupdater.SetRollingBlockHistory(gu, blocks)
+
+		gu.Recalculate(*cltest.Head(0))
+
+		ethClient.AssertExpectations(t)
+		config.AssertExpectations(t)
+	})
+
+	t.Run("handles unreasonably large gas prices (larger than a 64 bit int can hold)", func(t *testing.T) {
+		// Seems unlikely we will ever experience gas prices > 9 Petawei on mainnet (praying to the eth Gods 🙏)
+		// But other chains could easily use a different base of account
+		ethClient := new(mocks.Client)
+		config := new(gumocks.Config)
+
+		reasonablyHugeGasPrice := big.NewInt(0).Mul(big.NewInt(math.MaxInt64), big.NewInt(1000))
+
+		config.On("EthMaxGasPriceWei").Return(reasonablyHugeGasPrice)
+		config.On("EthMinGasPriceWei").Return(big.NewInt(10))
+		config.On("GasUpdaterTransactionPercentile").Return(uint16(50))
+		config.On("ChainID").Return(big.NewInt(0))
+
+		guIface := gasupdater.NewGasUpdater(ethClient, config)
+		gu := gasupdater.GasUpdaterToStruct(guIface)
+
+		unreasonablyHugeGasPrice := big.NewInt(0).Mul(big.NewInt(math.MaxInt64), big.NewInt(1000000))
+
+		b1Hash := cltest.NewHash()
+
+		blocks := []gasupdater.Block{
+			gasupdater.Block{
+				Number:     0,
+				Hash:       b1Hash,
+				ParentHash: common.Hash{},
+				Transactions: []gasupdater.Transaction{
+					gasupdater.Transaction{GasPrice: big.NewInt(50), GasLimit: 42},
+					gasupdater.Transaction{GasPrice: unreasonablyHugeGasPrice, GasLimit: 42},
+					gasupdater.Transaction{GasPrice: unreasonablyHugeGasPrice, GasLimit: 42},
+					gasupdater.Transaction{GasPrice: unreasonablyHugeGasPrice, GasLimit: 42},
+					gasupdater.Transaction{GasPrice: unreasonablyHugeGasPrice, GasLimit: 42},
+					gasupdater.Transaction{GasPrice: unreasonablyHugeGasPrice, GasLimit: 42},
+					gasupdater.Transaction{GasPrice: unreasonablyHugeGasPrice, GasLimit: 42},
+					gasupdater.Transaction{GasPrice: unreasonablyHugeGasPrice, GasLimit: 42},
+					gasupdater.Transaction{GasPrice: unreasonablyHugeGasPrice, GasLimit: 42},
+				},
+			},
+		}
+
+		config.On("SetEthGasPriceDefault", reasonablyHugeGasPrice).Return(nil)
+		gasupdater.SetRollingBlockHistory(gu, blocks)
+
+		gu.Recalculate(*cltest.Head(0))
+
+		ethClient.AssertExpectations(t)
+		config.AssertExpectations(t)
+	})
 }
 
 func TestGasUpdater_Block(t *testing.T) {
