+++ core/store/orm/config.go
@@ -82,6 +82,7 @@ type (
 		EthGasBumpWei                    *big.Int
 		EthGasPriceDefault               *big.Int
 		EthMaxGasPriceWei                *big.Int
+		EthMinGasPriceWei                *big.Int
 		EthFinalityDepth                 uint
 		EthHeadTrackerHistoryDepth       uint
 		EthHeadTrackerSamplingInterval   time.Duration
@@ -109,6 +110,7 @@ func init() {
 		EthGasBumpWei:                    big.NewInt(5000000000),    // 5 Gwei
 		EthGasPriceDefault:               big.NewInt(20000000000),   // 20 Gwei
 		EthMaxGasPriceWei:                big.NewInt(1500000000000), // 1.5 Twei
+		EthMinGasPriceWei:                big.NewInt(1000000000),    // 1 Gwei
 		EthFinalityDepth:                 50,
 		EthHeadTrackerHistoryDepth:       100,
 		EthHeadTrackerSamplingInterval:   1 * time.Second,
@@ -125,9 +127,23 @@ func init() {
 
 	// NOTE: There are probably other variables we can tweak for Kovan and other
 	// test chains, but requires more in-depth research on their consensus
-	// mechanisms. For now, mainnet defaults ought to be safe
+	// mechanisms. For now, mainnet defaults ought to be safe enough for testnet.
 	kovan := mainnet
 	kovan.HeadTimeBudget = 4 * time.Second
+
+	// xDai currently uses AuRa (like Parity) consensus so finality rules will be similar to parity
+	// See: https://www.poa.network/for-users/whitepaper/poadao-v1/proof-of-authority
+	// NOTE: xDai is planning to move to Honeybadger BFT which might have different finality guarantees
+	// https://www.xdaichain.com/for-validators/consensus/honeybadger-bft-consensus
+	// For worst case re-org depth on AuRa, assume 2n+2 (see: https://github.com/poanetwork/wiki/wiki/Aura-Consensus-Protocol-Audit)
+	// With xDai's current maximum of 19 validators then 40 blocks is the maximum possible re-org)
+	// The mainnet default of 50 blocks is ok here
+	xDai := mainnet
+	xDai.EthGasBumpThreshold = 8                      // mainnet * 2.8 ish (5s vs 13s block time)
+	xDai.EthGasPriceDefault = big.NewInt(1000000000)  // 1 Gwei
+	xDai.EthMinGasPriceWei = big.NewInt(1000000000)   // 1 Gwei is the minimum accepted by the validators (unless whitelisted)
+	xDai.EthMaxGasPriceWei = big.NewInt(500000000000) // 500 Gwei
+	xDai.HeadTimeBudget = 5 * time.Second
 
 	// BSC uses Clique consensus with ~3s block times
 	// Clique offers finality within (N/2)+1 blocks where N is number of signers
@@ -138,7 +154,7 @@ func init() {
 		EthGasBumpWei:                    big.NewInt(5000000000),   // 5 Gwei
 		EthGasPriceDefault:               big.NewInt(5000000000),   // 5 Gwei
 		EthMaxGasPriceWei:                big.NewInt(500000000000), // 500 Gwei
-		EthFinalityDepth:                 50,                       // Keeping this > 11 because it's not expensive and gives us a safety margin
+		EthFinalityDepth:                 50,                       // Keeping this >> 11 because it's not expensive and gives us a safety margin
 		EthHeadTrackerHistoryDepth:       100,
 		EthHeadTrackerSamplingInterval:   1 * time.Second,
 		EthBalanceMonitorBlockDelay:      2,
@@ -203,6 +219,8 @@ func init() {
 	ChainSpecificDefaults[56] = bscMainnet
 	ChainSpecificDefaults[128] = hecoMainnet
 	ChainSpecificDefaults[80001] = polygonMatic
+
+	ChainSpecificDefaults[100] = xDai
 }
 
 func chainSpecificConfig(c Config) ChainSpecificDefaultSet {
@@ -262,6 +280,13 @@ func (c *Config) Validate() error {
 			c.EthGasBumpPercent(),
 			ethCore.DefaultTxPoolConfig.PriceBump,
 		)
+	}
+
+	if c.EthMinGasPriceWei().Cmp(c.EthGasPriceDefault()) > 0 {
+		return errors.New("ETH_MIN_GAS_PRICE_WEI must be less than or equal to ETH_GAS_PRICE_DEFAULT")
+	}
+	if c.EthMaxGasPriceWei().Cmp(c.EthGasPriceDefault()) < 0 {
+		return errors.New("ETH_MAX_GAS_PRICE_WEI must be greater than or equal to ETH_GAS_PRICE_DEFAULT")
 	}
 
 	if c.EthHeadTrackerHistoryDepth() < c.EthFinalityDepth() {
@@ -613,7 +638,6 @@ func (c Config) EthGasBumpWei() *big.Int {
 // EthMaxGasPriceWei is the maximum amount in Wei that a transaction will be
 // bumped to before abandoning it and marking it as errored.
 func (c Config) EthMaxGasPriceWei() *big.Int {
-
 	str := c.viper.GetString(EnvVarName("EthMaxGasPriceWei"))
 	if str != "" {
 		n, err := parseBigInt(str)
@@ -635,6 +659,24 @@ func (c Config) EthMaxGasPriceWei() *big.Int {
 // 0 value disables
 func (c Config) EthMaxUnconfirmedTransactions() uint64 {
 	return c.getWithFallback("EthMaxUnconfirmedTransactions", parseUint64).(uint64)
+}
+
+// EthMinGasPriceWei is the minimum amount in Wei that a transaction may be priced.
+// Chainlink will never send a transaction priced below this amount.
+func (c Config) EthMinGasPriceWei() *big.Int {
+	str := c.viper.GetString(EnvVarName("EthMinGasPriceWei"))
+	if str != "" {
+		n, err := parseBigInt(str)
+		if err != nil {
+			logger.Errorw(
+				"Invalid value provided for EthMinGasPriceWei, falling back to default.",
+				"value", str,
+				"error", err)
+		} else {
+			return n.(*big.Int)
+		}
+	}
+	return chainSpecificConfig(c).EthMinGasPriceWei
 }
 
 // EthNonceAutoSync enables/disables running the NonceSyncer on application start
@@ -684,6 +726,14 @@ func (c Config) EthGasLimitMultiplier() float32 {
 
 // SetEthGasPriceDefault saves a runtime value for the default gas price for transactions
 func (c Config) SetEthGasPriceDefault(value *big.Int) error {
+	min := c.EthMinGasPriceWei()
+	max := c.EthMaxGasPriceWei()
+	if value.Cmp(min) < 0 {
+		return errors.Errorf("cannot set default gas price to %s, it is below the minimum allowed value of %s", value.String(), min.String())
+	}
+	if value.Cmp(max) > 0 {
+		return errors.Errorf("cannot set default gas price to %s, it is above the maximum allowed value of %s", value.String(), max.String())
+	}
 	if c.runtimeStore == nil {
 		return errors.New("No runtime store installed")
 	}
